requirements.txt
google-genai
playwright
mysql-connector-python
python-dotenv